Move dor command specs into command files
